# backend/go-gin-sqlx/cmd/api/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.LoginRequest:
    properties:
      password:
        maxLength: 100
        minLength: 6
        type: string
      username:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
  types.ResponseAPI:
    properties:
      data: {}
      debug:
        $ref: '#/definitions/types.ResponseAPIDebug'
      message:
        type: string
    type: object
  types.ResponseAPIDebug:
    properties:
      endTime:
        description: ISO8601 format for consistency with StartTime
        type: string
      error:
        type: string
      requestId:
        type: string
      runtimeMs:
        description: Runtime in milliseconds for better precision
        type: integer
      startTime:
        description: ISO8601 format, e.g., "2025-01-09T15:04:05Z07:00"
        type: string
      version:
        type: string
    type: object
host: localhost:9000
info:
  contact: {}
  description: API documentations for the BE.
  title: General API
  version: "1.0"
paths:
...
  /api/v1/auth/data:
    get:
      consumes:
      - application/json
      description: Get User Info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ResponseAPI'
      security:
      - BearerAuth: []
      summary: User Info
      tags:
      - Auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Get Access token
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ResponseAPI'
      summary: Auth
      tags:
      - Auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"